requirements/base.txt
# Base
asgiref==3.4.1
pytz==2021.3
sqlparse==0.4.2

# Django
Django==3.2.11

# Environment
django-environ==0.8.1

# Passwords security
argon2-cffi==21.3.0